feat: show data by category
